pyproject.toml
[tool.poetry]
name = "my-app"
version = "0.1.0"
description = ""
authors = ["Evgeni121 <darevblago121@mail.ru>"]
readme = "README.md"
packages = [{include = "smart_home"}]

[tool.poetry.dependencies]
python = "^3.10"
kivy = "^2.1.0"
kivymd = "^1.1.1"
django = "^4.1.7"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
pytz = "^2022.7.1"
sqlparse = "^0.4.3"
requests = "^2.28.2"
sqlalchemy = "^2.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
